requirements.txt
streamlit
pandas
requests
selenium
webdriver-manager
pdfplumber
python-dotenv